client_host/requirements.txt
numpy==1.21.6
opencv-python==4.8.1.78
deeplabcut-live==1.0.4
deeplabcut-live-gui==1.0
pandas==1.3.5
pyzmq==25.1.1
matplotlib
Pillow==9.5.0